+++ lib/components/organisms/calendar/utility.dart
@@ -1,4 +1,4 @@
-import 'package:pilll/components/organisms/calendar/calendar_band_model.dart';
+import 'package:pilll/components/organisms/calendar/band/calendar_band_model.dart';
 import 'package:pilll/domain/calendar/date_range.dart';
 import 'package:pilll/entity/menstruation.dart';
 import 'package:pilll/entity/pill_sheet.dart';
