+++ lib/domain/menstruation/menstruation_edit_page.dart
@@ -8,7 +8,7 @@ import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/components/page/discard_dialog.dart';
-import 'package:pilll/components/organisms/calendar/calendar.dart';
+import 'package:pilll/components/organisms/calendar/monthly/calendar.dart';
 import 'package:pilll/domain/menstruation/components/calendar/calendar_date_header.dart';
 import 'package:pilll/components/organisms/calendar/weekly/weekly_calendar.dart';
 import 'package:pilll/domain/calendar/monthly_calendar_state.dart';
